requirements.txt
PySide6
mysql-connector-python
pyinstaller
python-dotenv